Project2-AutoRAG/pyproject.toml
[project]
name = "autorag"
version = "0.1.0"
description = "Self-improving RAG system using the autoresearch pattern against Meta's CRAG benchmark"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "voyageai>=0.3.0",
    "openai>=1.60.0",
    "lancedb==0.30.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.2.0",
]
